+++ OUDS/Core/Components/Sources/Controls/Chip/Internal/Chip.swift
@@ -28,6 +28,8 @@ struct Chip: View {
 
     @Environment(\.theme) private var theme
 
+    // MARK: Enums
+
     enum IconPosition {
         case leading
         case trailing
@@ -35,8 +37,8 @@ struct Chip: View {
 
     enum Layout {
         case text(String)
-        case icon(Image, String)
-        case textAndIcon(text: String, icon: Image, iconPosition: IconPosition = .leading)
+        case icon(Image, String, renderingMode: Image.TemplateRenderingMode = .template)
+        case textAndIcon(text: String, icon: Image, iconPosition: IconPosition = .leading, renderingMode: Image.TemplateRenderingMode = .template)
     }
 
     // MARK: Initializers
@@ -79,7 +81,7 @@ struct Chip: View {
                 theme.chip.spacePaddingInlineIconNone
             case .icon:
                 theme.chip.spacePaddingInlineIcon
-            case let .textAndIcon(_, _, iconPosition):
+            case let .textAndIcon(_, _, iconPosition, _):
                 if iconPosition == .leading {
                     theme.chip.spacePaddingInlineIcon
                 } else {
@@ -95,7 +97,7 @@ struct Chip: View {
             theme.chip.spacePaddingInlineIconNone
         case .icon:
             theme.chip.spacePaddingInlineIcon
-        case let .textAndIcon(_, _, iconPosition):
+        case let .textAndIcon(_, _, iconPosition, _):
             if iconPosition == .trailing {
                 theme.chip.spacePaddingInlineIcon
             } else {
@@ -122,23 +124,23 @@ private struct ChipContent: View {
     var body: some View {
         Group {
             switch layout {
-            case let .icon(icon, accessibilityLabel):
-                ScaledIcon(icon: icon.renderingMode(.template),
+            case let .icon(icon, accessibilityLabel, renderingMode):
+                ScaledIcon(icon: icon.renderingMode(renderingMode),
                            size: theme.chip.sizeIcon)
                     .accessibilityLabel(accessibilityLabel)
             case let .text(text):
                 ChipText(text: text)
-            case let .textAndIcon(text, icon, iconPosition):
+            case let .textAndIcon(text, icon, iconPosition, renderingMode):
                 HStack(alignment: .center, spacing: theme.chip.spaceColumnGapIcon) {
                     if iconPosition == .leading {
-                        FixedIcon(icon: icon.resizable().renderingMode(.template),
+                        FixedIcon(icon: icon.resizable().renderingMode(renderingMode),
                                   size: theme.chip.sizeIcon)
                     }
 
                     ChipText(text: text)
 
                     if iconPosition == .trailing {
-                        FixedIcon(icon: icon.resizable().renderingMode(.template),
+                        FixedIcon(icon: icon.resizable().renderingMode(renderingMode),
                                   size: theme.chip.sizeIcon)
                     }
                 }